thinking on the go-git variant is bad too so it should just be removed
